refactored handler to take parameters so it can be called from other code
